backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["jonra1993 <jon_ra@hotmail.es>"]
package-mode = false

[tool.black]
line-length = 88
target-version = [ "py310", "py311" ]
exclude = "((.eggs | .git | .pytest_cache | build | dist))"

[tool.ruff]
line-length = 88
exclude = [".git", "__pycache__", ".mypy_cache", ".pytest_cache"]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    # "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "B904",
    "B006",    
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["alembic", "__pycache__"]

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.13.0"
prometheus_client = "^0.21.1"
redis = "^4.5.1"
minio = "^7.2.0"
celery = "^5.3.6"
celery-sqlalchemy-scheduler = "^0.3.0"
asyncpg = "^0.29.0"
fastapi = {extras = ["all"], version = "^0.110.0"}
sqlmodel = "0.0.16"
cryptography = "^41.0.7"
bcrypt = "^4.1.1"
SQLAlchemy-Utils = "^0.41.1"
SQLAlchemy = "^2.0.23"
fastapi-pagination = {extras = ["sqlalchemy"], version = "^0.12.21"}
fastapi-cache2 = "^0.2.1"
Pillow = "^10.1.0"
watchfiles = "^0.21.0"
asyncer = "0.0.5"
httpx = "^0.25.2"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
fastapi-async-sqlalchemy = "^0.6.0"
oso = "^0.27.0"
psycopg2-binary = "^2.9.5"
requests = "^2.31.0"
wheel = "^0.42.0"
setuptools = "^69.0.2"
fastapi-limiter = "^0.1.5"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}

[tool.poetry.group.dev.dependencies]
coverage = "^7.6.10"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
pytest-cov = "^6.0.0"
pydocstyle = "^6.3.0"
mypy = "^1.14.1"
gitchangelog = "^3.0.4"
mkdocs = ">=1.6.1"
mkdocs-material = "*"

[tool.poetry.group.test.dependencies]
black = "^24.10.0"
ruff = "^0.0.256"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
